Added quite a lot again

+++ FysioApp/Controllers/AppointmentsController.cs
@@ -1,5 +1,6 @@
 ﻿using ApplicationCore.Abstractions;
 using ApplicationCore.Entities;
+using ApplicationCore.Entities.ApiEntities;
 using ApplicationCore.Entities.ApplicationUsers;
 using ApplicationCore.Utility;
 using FysioApp.Models.ViewModels.AppointmentViewModels;
@@ -10,6 +11,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Net.Http;
 using System.Security.Claims;
 using System.Threading.Tasks;
 
@@ -23,8 +25,10 @@ namespace FysioApp.Controllers
         private readonly ITeacherRepository _teacherRepository;
         private readonly IPatientRepository _patientRepository;
 
+        private static readonly HttpClient client = new HttpClient();
+
         public AppointmentsController(
-            IAppointmentRepository appointmentRepository, 
+            IAppointmentRepository appointmentRepository,
             IIdentityUserRepository identityUserRepository,
             IStudentRepostitory studentRepostitory,
             IPatientRepository patientRepository,
@@ -55,7 +59,7 @@ namespace FysioApp.Controllers
         }
 
         //Get for Create
-        public IActionResult Create()
+        public async Task<IActionResult> Create()
         {
             var claimsIdentity = (ClaimsIdentity)this.User.Identity;
             var userId = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier).Value;
@@ -64,7 +68,7 @@ namespace FysioApp.Controllers
             {
                 Appointment = new Appointment()
                 {
-                    DateTime = DateTime.Now
+                    DateTime = DateTime.Now.AddHours(1)
                 },
                 Students = _studentRepository.GetStudents().ToList(),
                 Patients = _patientRepository.GetPatients().ToList()
@@ -125,38 +129,37 @@ namespace FysioApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(CreateAppointmentViewModel model)
         {
-
-            if (ModelState.IsValid)
-            {
-                if (model.Appointment.DateTime >= DateTime.Now)
-                {
-                    var appointment = new Appointment()
-                    {
-                        Description = model.Appointment.Description,
-                        DateTime = model.Appointment.DateTime,
-                        PatientId = model.Appointment.PatientId,
-                        StudentId = model.Appointment.StudentId,
-                        IsCancelled = false
-                    };
-
-                    _appointmentRepository.CreateAppointment(appointment);
-                    _appointmentRepository.Save();
-                    return RedirectToAction(nameof(Index));
-                } else
-                {
-                    ModelState.AddModelError(string.Empty, "Datum moet in de toekomst liggen");
-                }                   
-
-                
-            }
-
+            //Create a new viewmodel object, for when the form fails and everything needs t be reloaded into the view
             CreateAppointmentViewModel modelVM = new CreateAppointmentViewModel()
-            {                
+            {
                 Students = _studentRepository.GetStudents().ToList(),
                 Patients = _patientRepository.GetPatients().ToList(),
                 Appointment = model.Appointment
-
             };
+
+            if (ModelState.IsValid)
+            {
+                if (model.Appointment.DateTime < DateTime.Now)
+                {
+                    ModelState.AddModelError(string.Empty, "Datum en Tijd moet in de toekomst liggen.");
+                    return View(modelVM);
+                }
+
+                // create a new appointment object
+                Appointment appointment = new Appointment()
+                {
+                    Description = model.Appointment.Description,
+                    DateTime = model.Appointment.DateTime,
+                    PatientId = model.Appointment.PatientId,
+                    StudentId = model.Appointment.StudentId,
+                    IsCancelled = false,
+                };
+
+                _appointmentRepository.CreateAppointment(appointment); //create new appointment with object properties
+                _appointmentRepository.Save();                         //save to db
+                return RedirectToAction(nameof(Index));                //redirect to list of appointments
+            }
+
             return View(modelVM);
 
         }
@@ -166,6 +169,15 @@ namespace FysioApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(CreateAppointmentViewModel model)
         {
+            //Create a new viewmodel object, for when the form fails and everything needs t be reloaded into the view
+            CreateAppointmentViewModel modelVM = new CreateAppointmentViewModel()
+            {
+                Students = _studentRepository.GetStudents().ToList(),
+                Patients = _patientRepository.GetPatients().ToList(),
+                Appointment = model.Appointment
+
+            };
+
             if (ModelState.IsValid)
             {
                 Appointment appointmentFromDb = await _appointmentRepository.GetAppointment(model.Appointment.Id).FirstOrDefaultAsync();
@@ -173,10 +185,22 @@ namespace FysioApp.Controllers
                 {
                     return NotFound();
                 }
-                appointmentFromDb.Description = model.Appointment.Description;
+                
+                if (DateTime.Now.AddHours(24) > appointmentFromDb.DateTime) //check if appointment is within 24 hours
+                {
+                    ModelState.AddModelError(string.Empty, "U mag de afspraak niet meer wijzigen, deze vindt plaats binnen 24 uur."); // if not return with error
+                    return View(modelVM);
+                }
+
+                if (model.Appointment.DateTime < DateTime.Now) //check if datetime is in future
+                {
+                    ModelState.AddModelError(string.Empty, "Datum en Tijd moet in de toekomst liggen."); // if not return with error
+                    return View(modelVM);
+                }
+
+                //change most props
                 appointmentFromDb.DateTime = model.Appointment.DateTime;
-                appointmentFromDb.StudentId = model.Appointment.StudentId;
-                appointmentFromDb.PatientId = model.Appointment.PatientId;
+                appointmentFromDb.Description = model.Appointment.Description;  
 
                 _appointmentRepository.Save();
                 return RedirectToAction(nameof(Index));
@@ -190,8 +214,17 @@ namespace FysioApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Cancel(int id)
         {
-
             Appointment appointment = await _appointmentRepository.GetAppointment(id).FirstOrDefaultAsync();
+            if (appointment == null)
+            {
+                return NotFound();
+            }
+            if (DateTime.Now.AddHours(24) > appointment.DateTime) //check if appointment is within 24 hours
+            {
+                ModelState.AddModelError(string.Empty, "U mag de afspraak niet meer annuleren, deze vindt plaats binnen 24 uur."); // if not return with error
+                return View(appointment);
+            }
+            
             appointment.IsCancelled = true;
             _appointmentRepository.Save();
             return RedirectToAction(nameof(Index));
@@ -202,8 +235,18 @@ namespace FysioApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            Appointment appointment = await _appointmentRepository.GetAppointment(id).FirstOrDefaultAsync();
+            if (appointment == null)
+            {
+                return NotFound();
+            }
+            if (DateTime.Now.AddHours(24) > appointment.DateTime) //check if appointment is within 24 hours
+            {
+                ModelState.AddModelError(string.Empty, "U mag de afspraak niet meer verwijderen, deze vindt plaats binnen 24 uur."); // if not return with error
+                return View(appointment);
+            }
             _appointmentRepository.DeleteAppointment(id);
-            _appointmentRepository.Save();           
+            _appointmentRepository.Save();
             return RedirectToAction(nameof(Index));
         }
     }
